Reorg test files
