+++ src/cli/extendConfig.ts
@@ -1,20 +1,23 @@
 import path from 'path';
 import { writeFile, writeJsonFile } from './utils';
-import { computeFileContents } from './getChanges';
+import { computeFileContents } from './computeFileContents';
 import _ from 'lodash';
 import bluebird from 'bluebird';
 import fs from 'fs-extra';
 
 export async function extendConfig(cwd: string) {
-  const changes = await computeFileContents(cwd);
-  const iterableChanges = _.map(changes, (change, destinationPath) => ({
-    destinationPath,
-    change,
-  }));
+  const fileContents = await computeFileContents(cwd);
+  const iterableFileContents = _.map(
+    fileContents,
+    (fileContent, destinationPath) => ({
+      destinationPath,
+      fileContent,
+    }),
+  );
 
   return await bluebird.each(
-    iterableChanges,
-    async ({ destinationPath, change: { content, type } }) => {
+    iterableFileContents,
+    async ({ destinationPath, fileContent: { content, type } }) => {
       await fs.mkdirp(path.dirname(destinationPath));
 
       if (type === 'json') {
